Latest

+++ application/controllers/Add_material_request.php
@@ -40,7 +40,7 @@ class Add_material_request extends MY_Controller {
         $view_data['members_dropdown'] = $this->_get_team_members_dropdown();
         $view_data['status_dropdown'] = $this->_get_status_dropdown();
         $view_data["projects_dropdown"] = $this->_get_projects_dropdown_for_income_and_epxenses("expenses");
-
+        $view_data['login_id'] = $this->Users_model->login_user_id();
         $this->template->rander("add_material_request/index", $view_data);
     }
 
@@ -105,7 +105,6 @@ class Add_material_request extends MY_Controller {
         $id = $this->input->post('id');
         $test =$this->Add_material_request_model->get_list()->result();
         $pro_test = $test[0]->ids;
-        // $pro_test = str_pad($pro_test, 3, '0', STR_PAD_LEFT);
         $view_data['ids'] = $pro_test;
         $mr_id = intval($pro_test + 1);
         $data = array(
@@ -164,8 +163,45 @@ class Add_material_request extends MY_Controller {
         }
     }
 
+    function get_pdf() {
+        $mpdf = new \Mpdf\Mpdf();
+        
+        $model_info = $this->Add_material_request_model->get_one($this->input->post('id'));
+        $model_info->project_id = $model_info->project_id ? $model_info->project_id : $this->input->post('project_id');
+        $model_info->user_id = $model_info->user_id ? $model_info->user_id : $this->input->post('user_id');
+
+        $view_data['model_info'] = $model_info;
+        $view_data['location_add_dropdown'] = array("0" => "-") + $this->Location_categories_model->get_dropdown_list(array("title"));
+        $view_data['categories_add_dropdown'] = array("0" => "-") + $this->Expense_categories_model->get_dropdown_list(array("title"));
+        $view_data['projects_add_dropdown'] = array("0" => "-") + $this->Projects_model->get_dropdown_list(array("title"));
+        $team_members = $this->Users_model->get_all_where(array("deleted" => 0, "user_type" => "staff"))->result();
+        $members_add_dropdown = array();
+
+        foreach ($team_members as $team_member) {
+            $members_add_dropdown[$team_member->id] = $team_member->first_name . " " . $team_member->last_name;
+        }
+        $view_data['members_add_dropdown'] = array("0" => "-") + $members_add_dropdown;
+        $view_data['status_dropdown'] = $this->Status_mr_model->get_dropdown_list(array("title"));
+
+        // $options = array("date_request" => $date_request, "category_id" => $category_id, "location_id" => $location_id, "project_id" => $project_id, "user_id" => $user_id, "status_id" => $status_id);
+
+        if($this->Users_model->login_user_id() == 17){
+        $list_data = $this->Add_material_request_model->get_details_approved()->result();
+        }else{
+        $list_data = $this->Add_material_request_model->get_details()->result();
+        }
+        $view_data['list_data'] = $list_data;
+        
+        $this->load->view("add_material_request/get_pdf", $view_data);
+        
+        $html = $this->load->view('add_material_request/get_pdf',[],true);
+        // $mpdf->WriteHTML($html);
+        // $mpdf->Output(); // opens in browser
+    }
+
     //get the expnese list data
     function list_data() {
+        // $this->access_only_allowed_members();
         $date_request = $this->input->post('date_request');
         $category_id = $this->input->post('category_id');
         $location_id = $this->input->post('location_id');
@@ -176,7 +212,11 @@ class Add_material_request extends MY_Controller {
 
 
         $options = array("date_request" => $date_request, "category_id" => $category_id, "location_id" => $location_id, "project_id" => $project_id, "user_id" => $user_id, "status_id" => $status_id);
+        if($this->Users_model->login_user_id() == 17){
+        $list_data = $this->Add_material_request_model->get_details_approved($options)->result();
+        }else{
         $list_data = $this->Add_material_request_model->get_details($options)->result();
+        }
         $result = array();
         foreach ($list_data as $data) {
             $result[] = $this->_make_row($data);
@@ -203,6 +243,7 @@ public function view($id = 0, $result = null) {
         $view_data['status_warna'] = $status_warna;
         $model_item_info = $this->Add_item_model->get_one($this->input->post('id'));
         $view_data['model_item_info'] = $model_item_info;
+        $view_data['login_id'] = $this->Users_model->login_user_id();
         
         $data = array(
             'mr_no' => $model_info->mr_no,
@@ -212,10 +253,15 @@ public function view($id = 0, $result = null) {
       $query = $this->Add_material_request_model->upddata($data);
              redirect('/add_material_request/');      
       }
-        
-        // if($this->input->post('addmr')){
-        //         $query  = $this->Add_material_request_model->upddata();
-        // }
+      if($this->input->post('apprmr')){
+      $query = $this->Add_material_request_model->upddata_appr($data);
+             redirect('/add_material_request/');      
+      }
+      
+      if($this->input->post('rjcmr')){
+      $query = $this->Add_material_request_model->upddata_rjct($data);
+             redirect('/add_material_request/');      
+      } 
 
         $this->template->rander("add_material_request/view", $view_data);
 
@@ -300,6 +346,10 @@ private function _get_material_info_item($mr_no) {
 
 
    private function _make_row_item($data) {
+        $data1 = $this->Add_material_request_model->get_all_where(array('mr_no' => $data->mr_no))->row();
+        $status_created = $data1->status_id;
+        
+        
             $row_data = array(
             "MR2021-" . $data->mr_no,
             $data->item_name,
@@ -307,19 +357,23 @@ private function _get_material_info_item($mr_no) {
             $data->qty,
             $data->measurement,
             $data->notes,
-
         );
-        //$this->access_only_admin();
+        // var_dump($row_data);
+        if($status_created == 2){
         $row_data[] = modal_anchor(get_uri("add_material_request/add_item"), "<i class='fa fa-pencil'></i>", array("class" => "edit", "title" => lang('edit_item_material_request'), "data-post-id" => $data->id))
                 . js_anchor("<i class='fa fa-times fa-fw'></i>", array('title' => lang('delete_item_material_request'), "class" => "delete", "data-id" => $data->id, "data-action-url" => get_uri("add_material_request/delete_item"), "data-action" => "delete-confirmation"));
         return $row_data;
+        }
+        else{
+        $row_data[] = "<i class='fa fa-check' aria-hidden='true'></i>";
+        return $row_data;
+        }
     }
 
     //get a row of expnese list
     private function _row_data($id) {
-        // $custom_fields = $this->Custom_fields_model->get_available_fields_for_table("expenses", $this->login_user->is_admin, $this->login_user->user_type);
         $options = array("id" => $id);
-        $data = $this->Add_material_request_model->get_details($options)->row();
+        $data = $this->Add_material_request_model->get_details($options)->result();
         return $this->_make_row($data);
     }
 
@@ -328,9 +382,15 @@ private function _get_material_info_item($mr_no) {
         $status_warna = $data->status_mr_title;
         if ($status_warna == "Pending"){
             $status_warna = "<span class='label label-warning'>" . $status_warna . "</span>";
-        }elseif
+        }if
         ($status_warna == "Created"){
             $status_warna = "<span class='label label-primary'>" . $status_warna . "</span>";
+        }if
+        ($status_warna == "Approved"){
+            $status_warna = "<span class='label label-success'>" . $status_warna . "</span>";
+        }if
+        ($status_warna == "Rejected"){
+            $status_warna = "<span class='label label-danger'>" . $status_warna . "</span>";
         }
         $row_data = array(
             $data->date_request,
@@ -360,6 +420,7 @@ private function _get_material_info_item($mr_no) {
         $model_info = $this->Add_item_model->get_one($this->input->post('id'));
 
         $view_data['model_info'] = $model_info;
+        
         $this->load->view('add_material_request/add_item', $view_data);
     }
     
@@ -383,6 +444,7 @@ private function _get_material_info_item($mr_no) {
             "notes" => $this->input->post('notes'),
 
     );
+    
 
     $save_id = $this->Add_item_model->save($data, $id);
     if ($save_id) {
@@ -412,6 +474,7 @@ private function _get_material_info_item($mr_no) {
             "notes" => $this->input->post('notes'),
 
     );
+    
 
     $save_id = $this->Add_item_model->save($data, $id);
     if ($save_id) {
@@ -420,7 +483,7 @@ private function _get_material_info_item($mr_no) {
         echo json_encode(array("success" => false, 'message' => lang('error_occurred')));
     }
 }
-    
+
 }
 
 /* End of file expenses.php */
